% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{association}
\alias{association}
\title{Association}
\format{Each row of the data.frame contains two variables:
\describe{
 \item{V1}{The ENSEMBL Peptide ID of the first protein}
 \item{V2}{The ENSEMBL Peptide ID of the second protein}
}}
\usage{
association
}
\description{
A subset of the Homo sapiens protein network data from STRINGdb,
release 10.5. This subset
contains only associations of proteins of combined score greater than or
equal to 900.
}
\examples{
association

}
\seealso{
\link[transcriptogramer]{Hs900}
}
\author{
Diego Morais
}
\keyword{datasets}
