% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{clusterVisualization}
\alias{clusterVisualization}
\alias{clusterVisualization,Transcriptogram-method}
\title{Displays graphs of the differentially expressed clusters}
\usage{
clusterVisualization(object, maincomp = FALSE, connected = FALSE,
  host = "127.0.0.1", port = 9091, clusters = NULL)

\S4method{clusterVisualization}{Transcriptogram}(object, maincomp = FALSE,
  connected = FALSE, host = "127.0.0.1", port = 9091, clusters = NULL)
}
\arguments{
\item{object}{An object of class Transcriptogram.}

\item{maincomp}{Logical value, set as TRUE if you want to display only the main component of
each cluster, the default value for this argument is FALSE.}

\item{connected}{Logical value, set as TRUE if you want to display only connected nodes,
the default value for this argument is FALSE.}

\item{host}{The domain name of the machine that is running the RedeR XML-RPC
server.}

\item{port}{An integer specifying the port on which the XML-RPC server should
listen.}

\item{clusters}{An integer vector specifying the clusters to be
displayed, if NULL, all clusters will be displayed.}
}
\value{
This function returns an object of the RedPort Class.
}
\description{
This method uses the \pkg{RedeR} package to display graphs of the
differentially expressed clusters. If the DE slot has a column named Symbol,
its contents will be used as node alias.
}
\details{
RedeR package requirements: Java Runtime Environment (>= 6).
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
\dontrun{
transcriptogram <- transcriptogramStep1(transcriptogram, GSE9988, GPL570)
transcriptogram <- transcriptogramStep2(transcriptogram)
levels <- c(rep(FALSE, 3), rep(TRUE, 3))
transcriptogram <- differentiallyExpressed(transcriptogram, levels, 0.005,
DEsymbols)
rdp <- clusterVisualization(transcriptogram)
}

}
\seealso{
\link[transcriptogramer]{differentiallyExpressed},
\link[transcriptogramer]{transcriptogramPreprocess},
\link[transcriptogramer]{GSE9988},
\link[transcriptogramer]{GPL570},
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogramStep1},
\link[transcriptogramer]{transcriptogramStep2},
\link[RedeR]{RedPort}
}
\author{
Diego Morais
}
