% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{DE}
\alias{DE}
\alias{DE,Transcriptogram-method}
\alias{DE-method,DE}
\title{Get DE}
\usage{
DE(object)

\S4method{DE}{Transcriptogram}(object)
}
\arguments{
\item{object}{An object of class Transcriptogram.}
}
\value{
This method returns the content of the DE slot of an object of
class Transcriptogram.
}
\description{
Gets the content of the DE slot of an object of class Transcriptogram.
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
DE(transcriptogram)

}
\seealso{
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogramPreprocess}
}
\author{
Diego Morais
}
