% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{DEsymbols}
\alias{DEsymbols}
\title{Dictionary Gene2Symbol}
\format{An object of class \code{data.frame} with 129 rows and 2 columns.}
\usage{
DEsymbols
}
\description{
A mapping between ENSEMBL Peptide ID and Symbol (Gene Name) of a reduced set
of proteins.
}
\examples{
DEsymbols

}
\author{
Diego Morais
}
\keyword{datasets}
