% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{getSlot}
\alias{getSlot}
\alias{getSlot,Transcriptogram-method}
\title{Get the content of a slot}
\usage{
getSlot(object, slot)

\S4method{getSlot}{Transcriptogram}(object, slot)
}
\arguments{
\item{object}{An object of class Transcriptogram}

\item{slot}{A character specifying the slot}
}
\value{
This method returns the content of the specified slot of an object
of class Transcriptogram.
}
\description{
Get the content of a slot of an object of class Transcriptogram.
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
ord <- getSlot(transcriptogram, 'ordering')

}
\seealso{
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogramPreprocess},
\link[transcriptogramer]{Transcriptogram-class}
}
\author{
Diego Morais
}
