% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{GPL570}
\alias{GPL570}
\title{Dictionary Protein2Probe}
\format{Each row of the data.frame contains two variables:
\describe{
 \item{Protein}{The ENSEMBL Peptide ID}
 \item{Probe}{The probe identifier}
}}
\usage{
GPL570
}
\description{
A mapping between ENSEMBL Peptide ID and probe identifier, for the Homo
sapiens
and the platform GPL570, [HG-U133_Plus_2] Affymetrix Human Genome U133
Plus 2.0 Array.
}
\details{
This dataset was created to map the Homo sapiens proteins, from STRINGdb
release 11, of combined score greater than or equal to 700.
}
\examples{
GPL570

}
\seealso{
\link[transcriptogramer]{GSE9988}
}
\author{
Diego Morais
}
\keyword{datasets}
