% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Hs800}
\alias{Hs800}
\title{Ordered Homo sapiens proteins of combined score equal or above 800}
\format{An object of class \code{character} of length 13273.}
\usage{
Hs800
}
\description{
A character vector containing the Homo sapiens proteins, from STRINGdb
release 10.5,
of combined score equal or above 800.
}
\examples{
Hs800

}
\author{
Diego Morais
}
\keyword{datasets}
