% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{HsBPTerms}
\alias{HsBPTerms}
\title{Dictionary Gene2GO}
\format{An object of class \code{data.frame} with 107897 rows and 2 columns.}
\usage{
HsBPTerms
}
\description{
A mapping between ENSEMBL Peptide ID and Gene Ontology, biological process,
terms of a set of proteins.
}
\examples{
HsBPTerms

}
\author{
Diego Morais
}
\keyword{datasets}
