% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Mm700}
\alias{Mm700}
\title{Ordered Mus musculus proteins of combined score greater than or equal to 700}
\format{An object of class \code{character} of length 13921.}
\usage{
Mm700
}
\description{
A character vector containing the Mus musculus proteins, from STRINGdb
release 10.5,
of combined score greater than or equal to 700.
}
\examples{
Mm700

}
\author{
Diego Morais
}
\keyword{datasets}
