% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Mm800}
\alias{Mm800}
\title{Ordered Mus musculus proteins of combined score greater than or equal to 800}
\format{An object of class \code{character} of length 12166.}
\usage{
Mm800
}
\description{
A character vector containing the Mus musculus proteins, from STRINGdb
release 10.5,
of combined score greater than or equal to 800.
}
\details{
Generated by The Transcriptogramer V.1.0 for Windows. Input
arguments: isothermal steps - 100; Monte Carlo steps - 20000;
cooling factor - 0.5; alpha value - 1.0;
percentual energy for initial temperature - 0.0001. Final energy:
337846752.
}
\examples{
Mm800

}
\author{
Diego Morais
}
\keyword{datasets}
