% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{radius<-}
\alias{radius<-}
\alias{radius}
\alias{radius<-,Transcriptogram-method}
\alias{radius-method}
\alias{radius,Transcriptogram-method}
\title{Radius}
\usage{
radius(object) <- value

radius(object)

\S4method{radius}{Transcriptogram}(object) <- value

\S4method{radius}{Transcriptogram}(object)
}
\arguments{
\item{object}{An object of class Transcriptogram.}

\item{value}{An non-negative integer referring to the window
radius required for some methods.}
}
\value{
This method returns the content of the radius slot of an object of
class Transcriptogram.
}
\description{
Retrieve or set the content of the radius slot of an object of class
Transcriptogram.
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
radius(transcriptogram) <- 80
radius(transcriptogram)

}
\seealso{
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogramPreprocess},
\link[transcriptogramer:transcriptogramStep2-method]{transcriptogramStep2},
\link[transcriptogramer:orderingProperties-method]{orderingProperties}
}
\author{
Diego Morais
}
