% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Rn800}
\alias{Rn800}
\title{Ordered Rattus norvegicus proteins of combined score greater than or equal
to 800}
\format{An object of class \code{character} of length 12437.}
\usage{
Rn800
}
\description{
A character vector containing the Rattus norvegicus proteins, from STRINGdb
release 10.5,
of combined score greater than or equal to 800.
}
\examples{
Rn800

}
\author{
Diego Morais
}
\keyword{datasets}
