% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Rn900}
\alias{Rn900}
\title{Ordered Rattus norvegicus proteins of combined score greater than or equal
to 900}
\format{An object of class \code{character} of length 10788.}
\usage{
Rn900
}
\description{
A character vector containing the Rattus norvegicus proteins, from STRINGdb
release 11,
of combined score greater than or equal to 900.
}
\details{
Generated by The Transcriptogramer V.1.0 for Windows. Input
arguments: isothermal steps - 100; Monte Carlo steps - 20000;
cooling factor - 0.5; alpha value - 1.0;
percentual energy for initial temperature - 0.0001. Final energy:
222438518.
}
\examples{
Rn900

}
\author{
Diego Morais
}
\keyword{datasets}
