% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{setRadius}
\alias{setRadius}
\alias{setRadius,Transcriptogram-method}
\title{Changes the radius}
\usage{
setRadius(.Object, radius)

\S4method{setRadius}{Transcriptogram}(.Object, radius)
}
\arguments{
\item{.Object}{An object of class Transcriptogram}

\item{radius}{An integer, non negative, number referring to the window
radius required for some
methods}
}
\value{
This method changes the value of the radius slot of an object of
class Transcriptogram.
}
\description{
Changes the value of the radius slot of an object of class Transcriptogram.
}
\examples{
transcriptogram <- transcriptogram.preprocess(association, Hs900, 50)
transcriptogram <- setRadius(transcriptogram, 80)

}
\seealso{
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogram.preprocess},
\link[transcriptogramer]{transcriptogramStep2-method},
\link[transcriptogramer]{orderingProperties}
}
\author{
Diego Morais
}
