% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Transcriptogram-class}
\alias{Transcriptogram-class}
\title{Class Transcriptogram}
\description{
This S4 class includes methods to use expression values with ordered
proteins.
}
\section{Slots}{

\describe{
\item{\code{association}}{A data frame containing two columns, and rows containing
proteins names
that are connected}

\item{\code{ordering}}{A data frame containing two columns, the first one with
proteins names,
and the second containing its respective position}

\item{\code{transcriptogramS1}}{A data frame produced as the result of averaging
over all identifiers
related to the same protein}

\item{\code{transcriptogramS2}}{A data frame produced as the result of averaging
over the window}

\item{\code{radius}}{An integer, non negative, number referring to the window radius}

\item{\code{status}}{An integer used internally to check the status of the object}

\item{\code{DE}}{A data frame of differentially expressed proteins}
}}

\seealso{
\link[transcriptogramer]{transcriptogramPreprocess},
\link[transcriptogramer]{DE},
\link[transcriptogramer]{radius},
\link[transcriptogramer]{orderingProperties},
\link[transcriptogramer]{connectivityProperties},
\link[transcriptogramer]{transcriptogramStep1},
\link[transcriptogramer]{transcriptogramStep2},
\link[transcriptogramer]{differentiallyExpressed},
\link[transcriptogramer]{clusterVisualization},
\link[transcriptogramer]{clusterEnrichment}
}
\author{
Diego Morais
}
