% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{transcriptogramer-package}
\alias{transcriptogramer}
\alias{transcriptogramer-package}
\title{Transcriptional analysis based on transcriptograms}
\description{
R package for transcriptional analysis based on transcriptograms, a method
to analyze transcriptomes that projects expression values on a set of
ordered proteins, arranged such that the probability that gene products
participate in the same metabolic pathway exponentially decreases with the
increase of the distance between two proteins of the ordering.
Transcriptograms are, hence, genome wide gene expression profiles that
provide a global view for the cellular metabolism, while indicating gene
sets whose expression are altered.
}
\details{
Performs  topological  analysis,  differential  expression,  and  Gene
Ontology enrichment analysis.

See the package at \url{https://doi.org/doi:10.18129/B9.bioc.transcriptogramer}
for complete documentation.
}
\author{
\strong{Maintainer}: Diego Morais \email{vinx@ufrn.edu.br}

Authors:
\itemize{
  \item Rodrigo Dalmolin \email{rodrigo.dalmolin@imd.ufrn.br}
}

}
