% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{transcriptogramStep1}
\alias{transcriptogramStep1}
\alias{transcriptogramStep1,Transcriptogram-method}
\title{Calculates the average of the expression values related to the same protein}
\usage{
transcriptogramStep1(.Object, expression, dictionary)

\S4method{transcriptogramStep1}{Transcriptogram}(.Object, expression,
  dictionary)
}
\arguments{
\item{.Object}{An object of class Transcriptogram}

\item{expression}{A matrix, or data frame, containing normalized expression
values from samples of microarrays
or RNA-Seq}

\item{dictionary}{A matrix, or data frame, containing two columns, the first
column must contains the
ENSEMBL Peptide ID, and the second column must contains values that appear
as rownames in \code{expression},
in order to recognize the ENSEMBL Peptide ID of the other column}
}
\value{
This method creates a data frame to feed the transcriptogramS1
slot of an object
of class Transcriptogram. Each row of the data frame contains: an ENSEMBL
Peptide ID, its
respective position in the ordering and the mean of the expression values
of the identifiers
related to the same protein.
}
\description{
For each transcriptome sample, this method assigns to each protein the
average of the expression values of all the identifiers related to
it. It is necessary a
\code{dictionary} to map the identifiers to proteins.
}
\examples{
transcriptogram <- transcriptogram.preprocess(association, Hs900)
\dontrun{
transcriptogram <- transcriptogramStep1(transcriptogram, GSE9988, GPL570)
}

}
\references{
da Silva, S.R.M., Perrone, G.C., Dinis, J.M. and de Almeida, R.M.C.,
Transcriptograms: Reproducibility enhancement and differential expression
of non predefined functional gene sets in human genome. BMC Genomics, 15,
1181 (2014).

Rybarczyk-Filho, J.L., Castro, M.A.A., Dalmolin, R.J, Moreira, J.C.F.,
Brunnet, L.G. and de Almeida, R.M.C., Towards a genome-wide transcriptogram:
the Saccharomyces cerevisiae case. Nucleic Acids Res., 39, 3005-3016 (2011).
PMID:21169199
}
\seealso{
\link[transcriptogramer]{transcriptogram.preprocess}
\link[transcriptogramer]{GSE9988}
\link[transcriptogramer]{GPL570}
\link[transcriptogramer]{Hs900}
\link[transcriptogramer]{association}
}
\author{
Diego Morais
}
