% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{transcriptogramStep2}
\alias{transcriptogramStep2}
\alias{transcriptogramStep2,Transcriptogram-method}
\alias{transcriptogramStep2-method}
\title{Calculates the average of the expression values using a sliding window}
\usage{
transcriptogramStep2(object, nCores = 1L)

\S4method{transcriptogramStep2}{Transcriptogram}(object, nCores = 1L)
}
\arguments{
\item{object}{An object of class Transcriptogram.}

\item{nCores}{An integer number, referring to the number of processing cores
to be used; or a logical value, TRUE indicating that all processing cores
should be used, and FALSE indicating the use of just one processing core.
The default value of this argument is 1.}
}
\value{
This method creates a data.frame to feed the transcriptogramS2
slot of an object
of class Transcriptogram. Each row of the data.frame contains: the ENSEMBL
Peptide ID used as center of the
window, its position on the ordering, and the mean of the expression values
of the window.
}
\description{
To each position of the ordering, this method assigns a value
equal to the average of the expression values inside a window, region of n
(radius * 2 + 1) proteins
centered at a protein. The window considers periodic boundary conditions to
deal
with proteins near the ends of the ordering.
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
\dontrun{
transcriptogram <- transcriptogramStep1(transcriptogram, GSE9988, GPL570)
transcriptogram <- transcriptogramStep2(transcriptogram)
}

}
\references{
da Silva, S.R.M., Perrone, G.C., Dinis, J.M. and de Almeida, R.M.C.,
Transcriptograms: Reproducibility enhancement and differential expression
of non predefined functional gene sets in human genome. BMC Genomics, 15,
1181 (2014).

Rybarczyk-Filho, J.L., Castro, M.A.A., Dalmolin, R.J, Moreira, J.C.F.,
Brunnet, L.G. and de Almeida, R.M.C., Towards a genome-wide transcriptogram:
the Saccharomyces cerevisiae case. Nucleic Acids Res., 39, 3005-3016 (2011).
PMID:21169199
}
\seealso{
\link[transcriptogramer]{transcriptogramPreprocess},
\link[transcriptogramer]{GSE9988},
\link[transcriptogramer]{GPL570},
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer:transcriptogramStep1-method]{transcriptogramStep1}
}
\author{
Diego Morais
}
