% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{DE}
\alias{DE}
\alias{DE,Transcriptogram-method}
\alias{DE-method}
\title{Get DE}
\usage{
DE(object)

\S4method{DE}{Transcriptogram}(object)
}
\arguments{
\item{object}{An object of class Transcriptogram.}
}
\value{
This method returns the content of the DE slot of an object of
class Transcriptogram.
}
\description{
Gets the content of the DE slot of an object of class Transcriptogram.
}
\examples{
transcriptogram <- transcriptogramPreprocess(association, Hs900, 50)
\dontrun{
transcriptogram <- transcriptogramStep1(transcriptogram, GSE9988, GPL570)
transcriptogram <- transcriptogramStep2(transcriptogram)
levels <- c(rep(FALSE, 3), rep(TRUE, 3))
transcriptogram <- differentiallyExpressed(transcriptogram, levels, 0.01)
DE(transcriptogram)
}

}
\seealso{
\link[transcriptogramer]{Hs900},
\link[transcriptogramer]{association},
\link[transcriptogramer]{transcriptogramPreprocess}
}
\author{
Diego Morais
}
