% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{DEsymbols}
\alias{DEsymbols}
\title{Dictionary Gene2Symbol}
\format{Each row of the data.frame contains two variables:
\describe{
 \item{ensembl_peptide_id}{The ENSEMBL Peptide ID}
 \item{external_gene_name}{The Gene Name}
}}
\usage{
DEsymbols
}
\description{
A mapping between ENSEMBL Peptide ID and Symbol (Gene Name) of a reduced set
of proteins.
}
\details{
This dataset was created to map the differentially expressed
proteins of the vignette example.
}
\examples{
DEsymbols

}
\author{
Diego Morais
}
\keyword{datasets}
