% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{HsBPTerms}
\alias{HsBPTerms}
\title{Dictionary Protein2GO}
\format{Each row of the data.frame contains two variables:
\describe{
 \item{ensembl_peptide_id}{The ENSEMBL Peptide ID}
 \item{go_id}{The Gene Ontology ID}
}}
\usage{
HsBPTerms
}
\description{
A mapping between ENSEMBL Peptide ID and Gene Ontology, biological process,
terms of a set of proteins.
}
\details{
This dataset was created to map the Homo sapiens proteins
that appear in the slot transcriptogramS2
of the vignette example.
}
\examples{
HsBPTerms

}
\author{
Diego Morais
}
\keyword{datasets}
