% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Rn700}
\alias{Rn700}
\title{Ordered Rattus norvegicus proteins of combined score equal or above 700}
\format{An object of class \code{character} of length 14285.}
\usage{
Rn700
}
\description{
A character vector containing the Rattus norvegicus proteins, from STRINGdb
release 10.5,
of combined score equal or above 700.
}
\examples{
Rn700

}
\author{
Diego Morais
}
\keyword{datasets}
