% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Rn900}
\alias{Rn900}
\title{Ordered Rattus norvegicus proteins of combined score equal or above 900}
\format{An object of class \code{character} of length 9747.}
\usage{
Rn900
}
\description{
A character vector containing the Rattus norvegicus proteins, from STRINGdb
release 10.5,
of combined score equal or above 900.
}
\examples{
Rn900

}
\author{
Diego Morais
}
\keyword{datasets}
