% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Sc700}
\alias{Sc700}
\title{Ordered Saccharomyces cerevisiae proteins of combined score greater than or
equal to 700}
\format{An object of class \code{character} of length 5586.}
\usage{
Sc700
}
\description{
A character vector containing the Saccharomyces cerevisiae proteins, from
STRINGdb release 10.5,
of combined score greater than or equal to 700.
}
\details{
Generated by The Transcriptogramer V.1.0 for Windows. Input
arguments: isothermal steps - 100; Monte Carlo steps - 20000;
cooling factor - 0.5; alpha value - 1.0;
percentual energy for initial temperature - 0.0001. Final energy:
141868972.
}
\examples{
Sc700

}
\author{
Diego Morais
}
\keyword{datasets}
