% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{Sc700}
\alias{Sc700}
\title{Ordered Saccharomyces cerevisiae proteins of combined score greater than or
equal to 700}
\format{An object of class \code{character} of length 5586.}
\usage{
Sc700
}
\description{
A character vector containing the Saccharomyces cerevisiae proteins, from
STRINGdb release 10.5,
of combined score greater than or equal to 700.
}
\examples{
Sc700

}
\author{
Diego Morais
}
\keyword{datasets}
